Added Factorial functionality to this project.
This will create a "partial" factorial of a given range of numbers. For a full
factorial, have the start point be "1". Included is a quick unit test in C with
some quick values checked.

I need to switch to SWIG and make it so that I can get everything working and
test using Python. Perhaps tomorrow when my laptop isn't about to die because I
forgot my powercord.