Updated the README file requirements, change log, and file manifest.
I added the requirement for compilation to be in a C99 compatible environment.
I also listed the new files that have been incorporated into the program in the
file manifest. Finally, I updated the change log to match these new features.

This branch are ready to merge into master and master tagged to v0.0.2.