Added new big number scheme.
Needed to be able to handle huge numbers in C. This set of functions will allow
for a set number of digits to be computed then the basic math functions for
them to be computed. The digits are stored in a char array to try and conserve
memory. Multiplication and Division are working off basic addition and
subtraction however division is still the slowest operation.

Time to work on redoing factorials and combinations again.